more examples ported to surface syntax
