+++ hydroflow/deprecated_examples/covid_tracing/people.rs
