+++ hydroflow/deprecated_examples/three_clique/README.md
@@ -1,0 +1,8 @@
+# Clique of size 3
+
+To run:
+```
+cargo run -p hydroflow --example three_clique
+```
+
+Adding the `-- --graph <graph_type>` flag to the end of the command line above will print out a node-and-edge diagram of the program. Supported values for `<graph_type>` include [mermaid](https://mermaid-js.github.io/) and [dot](https://graphviz.org/doc/info/lang.html).
