--- hydroflow/examples/surface_syntax/README.md
@@ -1,1 +1,0 @@
-# Surface Syntax
