+++ hydroflow/examples/three_clique/main.rs
@@ -1,6 +1,5 @@
 use clap::{ArgEnum, Parser};
-use hydroflow::builder::prelude::*;
-use hydroflow::scheduled::handoff::VecHandoff;
+use hydroflow::hydroflow_syntax;
 
 //This example detects size three cliques in a graph. Size three cliques are also known as triangles.
 //The equivalent datalog program would be Triangle(x,y,z) := Edge(x,y), Edge(y,z), Edge(z,x)
@@ -13,68 +12,71 @@ enum GraphType {
 #[derive(Parser, Debug)]
 struct Opts {
     #[clap(arg_enum, long)]
-    graph: GraphType,
+    graph: Option<GraphType>,
 }
 pub fn main() {
     let opts = Opts::parse();
-    let mut builder = HydroflowBuilder::default();
+    // An edge in the input data = a pair of `usize` vertex IDs.
+    let (edges_send, edges_recv) = hydroflow::util::unbounded_channel::<(usize, usize)>();
 
-    let (send_edges, recv_edges) =
-        builder.add_channel_input::<_, _, VecHandoff<(usize, usize)>>("edge input");
-    let (send_a, recv_a) = builder.make_edge::<_, VecHandoff<(usize, usize)>, _>("handoff_a");
-    let (send_b, recv_b) = builder.make_edge::<_, VecHandoff<(usize, usize)>, _>("handoff_b");
-    let (send_c, recv_c) = builder.make_edge::<_, VecHandoff<(usize, usize)>, _>("handoff_c");
+    let mut df = hydroflow_syntax! {
+        edges = recv_stream(edges_recv) -> tee();
 
-    builder.add_subgraph(
-        "teeing",
-        recv_edges.flatten().pull_to_push().map(Some).tee(
-            builder.start_tee().push_to(send_a),
-            builder.start_tee().tee(
-                builder.start_tee().push_to(send_b),
-                builder.start_tee().push_to(send_c),
-            ),
-        ),
-    );
+        // set up the two joins
+        // edge_pairs((z,x), y) :- edges(x,y), edges(y,z)
+        edge_pairs = join() -> map(|(y, (x,z))| ((z,x), y)); //Here we have found all paths from x to z that go through y. Now we need to find edges that connect z back to x.
+        // triangle(x,y,z) :- edge_pairs((z,x), y), edges(z, x)
+        triangle = join() -> map(|((z,x), (y, ()))| (x, y, z));
 
-    builder.add_subgraph(
-        "joining",
-        recv_a
-            .flatten()
-            .map(|(x, y)| (y, x))
-            .join(recv_b.flatten())
-            .map(|(y, (x, z))| ((z, x), y)) //Here we have found all paths from x to z that go through y. Now we need to find edges that connect z back to x.
-            .join(recv_c.flatten().map(|(z, x)| ((z, x), ())))
-            .inspect(|&v| println!("three_clique found: {:?}", v))
-            .pull_to_push()
-            .for_each(|_| {}),
-    );
+        // wire the inputs to the joins
+        edges[0] -> map(|(y,z)| (z,y)) -> [0]edge_pairs;
+        edges[1] -> [1]edge_pairs;
+        edge_pairs -> map(|((z,x), y)| ((z, x), y)) -> [0]triangle;
+        edges[2] -> map(|(z,x)| ((z,x), ())) -> [1]triangle;
 
-    let mut hf = builder.build();
-    match opts.graph {
-        GraphType::Mermaid => {
-            println!("{}", hf.generate_mermaid())
-        }
-        GraphType::Dot => {
-            println!("{}", hf.generate_dot())
-        }
-        GraphType::Json => {
-            println!("{}", hf.generate_json())
+        // post-process: sort fields of each tuple by node ID
+        triangle -> map(|(x, y, z)| {
+            let mut v = vec![x, y, z];
+            v.sort();
+            (v[0], v[1], v[2])
+        }) -> for_each(|e| println!("three_clique found: {:?}", e))
+    };
+
+    if let Some(graph) = opts.graph {
+        match graph {
+            GraphType::Mermaid => {
+                println!("{}", df.generate_mermaid())
+            }
+            GraphType::Dot => {
+                println!("{}", df.generate_dot())
+            }
+            GraphType::Json => {
+                println!("{}", df.generate_json())
+            }
         }
     }
+    df.run_available();
 
     println!("A");
 
-    send_edges.give(Some((5, 10)));
-    send_edges.give(Some((0, 3)));
-    send_edges.give(Some((3, 6)));
-    send_edges.flush();
-    hf.run_available();
+    edges_send.send((5, 10)).unwrap();
+    edges_send.send((0, 3)).unwrap();
+    edges_send.send((3, 6)).unwrap();
+    df.run_available();
 
     println!("B");
 
-    send_edges.give(Some((6, 5)));
-    send_edges.give(Some((6, 0))); //Creates a size three clique (triangle)
-    send_edges.give(Some((10, 6))); //Creates a size three clique (triangle)
-    send_edges.flush();
-    hf.run_available();
+    edges_send.send((6, 5)).unwrap();
+    edges_send.send((6, 0)).unwrap();
+    edges_send.send((10, 6)).unwrap();
+    df.run_available();
+
+    // A
+    // B
+    // three_clique found: (0, 3, 6)
+    // three_clique found: (5, 6, 10)
+    // three_clique found: (0, 3, 6)
+    // three_clique found: (5, 6, 10)
+    // three_clique found: (0, 3, 6)
+    // three_clique found: (5, 6, 10)
 }
